project/requirements.txt
requests==2.22.0
beautifulsoup4==4.6.3
LINE-BOT-SDK==1.16.0
pandas==1.0.1
Flask==1.1.1
gunicorn==20.0.4
lxml==4.5.2
SQLAlchemy==1.3.11
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
psycopg2==2.8.4
python-dateutil==2.8.1
gspread==3.6.0